fix e2e layout + automatic scenarios collection
