+++ Cargo.toml
@@ -79,6 +79,7 @@ reqwest = { version = "0.13.0" }
 rustls = "0.23.0"
 serde = "1.0.145"
 serde_json = "1.0.86"
+toml = "0.8"
 siphasher = "1"
 socket2 = "0.6.4"
 smallvec = "1.15.1"
