+++ yellowstone-grpc-e2e-test/Cargo.toml
@@ -10,8 +10,10 @@ publish = false
 
 [dependencies]
 anyhow = "1"
+futures = { workspace = true }
 clap = { workspace = true, features = ["derive"] }
 dotenvy = { workspace = true }
+humantime-serde = { workspace = true }
 inventory = { workspace = true }
 log = { workspace = true }
 env_logger = { workspace = true }
@@ -26,6 +28,8 @@ yellowstone-block-machine = { workspace = true }
 yellowstone-grpc-client = { workspace = true, features = ["test-tools"] }
 yellowstone-grpc-geyser = { workspace = true }
 
+serde = { workspace = true, features = ["derive"] }
+toml = { workspace = true }
 solana-signature = { workspace = true }
 solana-pubkey = { workspace = true }
 solana-commitment-config = { workspace = true }
