+++ yellowstone-grpc-e2e-test/src/bin/yellowstone-e2e.rs
@@ -1,70 +1,38 @@
 use {
     anyhow::{Context, Result},
-    clap::{Parser, Subcommand, ValueEnum},
+    clap::{Parser, Subcommand},
     std::{collections::HashMap, env, io::Write, path::PathBuf, process::ExitCode},
     tokio::time::{self, Duration, MissedTickBehavior},
-    yellowstone_grpc_intg_test::scenarios::{
-        any_commitment_level_of_subscription_should_return_all_possible_values, init_log,
-        it_should_subscribe_to_all_transaction_include_token_ata_to_an_owner,
-        it_should_support_replay, it_should_verifies_geyser_event_ordering_is_correct,
-        it_should_verify_replay_ordering_matches_live_path, scenario_description,
-        slot_status_should_have_parent, subscribe_should_filter_accounts,
-        subscribe_should_only_returns_sysvarclock_account,
-        subscribe_should_receive_block_where_sysvarclock1111_account_has_been_updated,
-        subscribe_should_receive_full_blocks, subscribe_should_receive_no_slot_duplicates,
-        test_subscribe_deshred, RunConfig,
+    yellowstone_grpc_intg_test::{
+        config::Config,
+        scenarios::{init_log, RunConfig, Scenario},
     },
 };
-
-#[derive(Debug, Clone, ValueEnum)]
-enum Scenario {
-    SysvarAccount,
-    SysvarBlock,
-    FullBlocks,
-    Replay,
-    Deshred,
-    AnyCommitment,
-    TokenOwnerBalanceChanged,
-    Ordering,
-    FilterAccounts,
-    SlotDuplicate,
-    ReplayOrdering,
-    SlotStatusParentPresent,
-}
-
-impl Scenario {
-    const fn name(&self) -> &'static str {
-        match self {
-            Self::SysvarAccount => "sysvar-account",
-            Self::SysvarBlock => "sysvar-block",
-            Self::FullBlocks => "full-blocks",
-            Self::Replay => "replay",
-            Self::Deshred => "deshred",
-            Self::AnyCommitment => "any-commitment",
-            Self::TokenOwnerBalanceChanged => "token-owner-balance-changed",
-            Self::Ordering => "event-ordering",
-            Self::FilterAccounts => "filter-accounts",
-            Self::SlotDuplicate => "slot-duplicate",
-            Self::ReplayOrdering => "replay-ordering",
-            Self::SlotStatusParentPresent => "slot-status-parent-present",
-        }
-    }
-
-    fn description(&self) -> &'static str {
-        scenario_description(self.name()).unwrap_or("No description available")
-    }
-}
 
 #[derive(Debug, Subcommand)]
 enum Commands {
     /// List all available e2e subscriber scenarios.
-    List,
+    List {
+        /// Only show scenarios that carry this tag (repeatable; any match passes).
+        #[arg(long = "tag", value_name = "TAG")]
+        tags: Vec<String>,
+        /// Only show scenarios from this module (e.g. `default`, `extra`).
+        #[arg(long = "module", value_name = "MODULE", default_value = "default")]
+        module: String,
+    },
     /// Run all e2e subscriber scenarios.
-    All,
+    All {
+        /// Only run scenarios that carry this tag (repeatable; any match passes).
+        #[arg(long = "tag", value_name = "TAG")]
+        tags: Vec<String>,
+        /// Only run scenarios from this module (e.g. `default`, `extra`).
+        #[arg(long = "module", value_name = "MODULE", default_value = "default")]
+        module: String,
+    },
     /// Run one specific subscriber scenario.
     Run {
-        #[arg(value_enum)]
-        scenario: Scenario,
+        #[arg(value_name = "SCENARIO")]
+        scenario: String,
     },
 }
 
@@ -90,6 +58,10 @@ struct Cli {
     /// Dotenv file path override. If set, only this file is loaded.
     #[arg(long, value_name = "PATH")]
     dotenv: Option<PathBuf>,
+
+    /// TOML config file for scenario-specific parameters.
+    #[arg(long, value_name = "PATH")]
+    config_file: Option<PathBuf>,
 }
 
 fn load_dotenv(dotenv_path_override: Option<&PathBuf>) -> HashMap<String, String> {
@@ -182,36 +154,33 @@ fn resolve_dial(cli: &Cli, dotenv_values: &HashMap<String, String>) -> Option<St
         .or_else(|| env::var("YELLOWSTONE_GRPC_DIAL").ok())
 }
 
-async fn run_scenario(scenario: &Scenario, config: &RunConfig) -> Result<()> {
-    let mut result = Box::pin(async {
-        match scenario {
-            Scenario::SysvarAccount => {
-                subscribe_should_only_returns_sysvarclock_account(config).await
-            }
-            Scenario::SysvarBlock => {
-                subscribe_should_receive_block_where_sysvarclock1111_account_has_been_updated(
-                    config,
-                )
-                .await
-            }
-            Scenario::FullBlocks => subscribe_should_receive_full_blocks(config).await,
-            Scenario::Replay => it_should_support_replay(config).await,
-            Scenario::Deshred => test_subscribe_deshred(config).await,
-            Scenario::AnyCommitment => {
-                any_commitment_level_of_subscription_should_return_all_possible_values(config).await
-            }
-            Scenario::TokenOwnerBalanceChanged => {
-                it_should_subscribe_to_all_transaction_include_token_ata_to_an_owner(config).await
-            }
-            Scenario::Ordering => it_should_verifies_geyser_event_ordering_is_correct(config).await,
-            Scenario::FilterAccounts => subscribe_should_filter_accounts(config).await,
-            Scenario::SlotDuplicate => subscribe_should_receive_no_slot_duplicates(config).await,
-            Scenario::ReplayOrdering => {
-                it_should_verify_replay_ordering_matches_live_path(config).await
-            }
-            Scenario::SlotStatusParentPresent => slot_status_should_have_parent(config).await,
-        }
-    });
+fn matches_tags(scenario: &Scenario, tags: &[String]) -> bool {
+    tags.is_empty() || tags.iter().any(|t| scenario.tags.contains(&t.as_str()))
+}
+
+fn matches_module(scenario: &Scenario, module: &str) -> bool {
+    scenario.module == module || scenario.module.ends_with(&format!("::{module}"))
+}
+
+fn find_scenario(name: &str) -> Result<&'static Scenario> {
+    inventory::iter::<Scenario>
+        .into_iter()
+        .find(|s| s.name == name)
+        .ok_or_else(|| {
+            let available: Vec<_> = inventory::iter::<Scenario>
+                .into_iter()
+                .map(|s| s.name)
+                .collect();
+            anyhow::anyhow!(
+                "unknown scenario '{}'; available: {}",
+                name,
+                available.join(", ")
+            )
+        })
+}
+
+async fn run_scenario(scenario: &'static Scenario, config: &RunConfig) -> Result<()> {
+    let mut result = Box::pin((scenario.run)(config));
     let mut interval = time::interval(Duration::from_millis(120));
     interval.set_missed_tick_behavior(MissedTickBehavior::Delay);
 
@@ -224,12 +193,12 @@ async fn run_scenario(scenario: &Scenario, config: &RunConfig) -> Result<()> {
             res = &mut result => {
                 write!(stdout, "\r\x1b[2K")?;
                 if res.is_ok() {
-                    writeln!(stdout, "✅ scenario '{}' passed", scenario.name())?;
+                    writeln!(stdout, "✅ scenario '{}' passed", scenario.name)?;
                 } else {
                     writeln!(
                         stdout,
                         "❌ scenario '{}' failed: {:#}",
-                        scenario.name(),
+                        scenario.name,
                         res.as_ref().expect_err("error should be present on failure")
                     )?;
                 }
@@ -241,7 +210,7 @@ async fn run_scenario(scenario: &Scenario, config: &RunConfig) -> Result<()> {
                     stdout,
                     "\r{} running scenario '{}'...",
                     frames[frame_index],
-                    scenario.name()
+                    scenario.name
                 )?;
                 stdout.flush()?;
                 frame_index = (frame_index + 1) % frames.len();
@@ -253,20 +222,39 @@ async fn run_scenario(scenario: &Scenario, config: &RunConfig) -> Result<()> {
 async fn run(cli: Cli) -> Result<()> {
     init_log();
 
-    if let Commands::List = cli.command {
-        let scenarios = [
-            Scenario::SysvarAccount,
-            Scenario::SysvarBlock,
-            Scenario::FullBlocks,
-            Scenario::Replay,
-            Scenario::Deshred,
-            Scenario::AnyCommitment,
-            Scenario::TokenOwnerBalanceChanged,
-            Scenario::ReplayOrdering,
-        ];
+    if let Commands::List {
+        ref tags,
+        ref module,
+    } = cli.command
+    {
+        let scenarios: Vec<&Scenario> = inventory::iter::<Scenario>
+            .into_iter()
+            .filter(|s| matches_tags(s, tags) && matches_module(s, module))
+            .collect();
+
+        if scenarios.is_empty() {
+            println!("No scenarios match the specified tags.");
+            return Ok(());
+        }
+
+        let name_w = scenarios.iter().map(|s| s.name.len()).max().unwrap_or(4);
+        let tags_w = scenarios
+            .iter()
+            .map(|s| s.tags.join(", ").len())
+            .max()
+            .unwrap_or(4)
+            .max(4);
+
+        println!("{:<name_w$}  {:<tags_w$}  DESCRIPTION", "NAME", "TAGS",);
+        println!("{}", "─".repeat(name_w + 2 + tags_w + 2 + 40));
 
         for scenario in scenarios {
-            println!("{} - {}", scenario.name(), scenario.description());
+            println!(
+                "{:<name_w$}  {:<tags_w$}  {}",
+                scenario.name,
+                scenario.tags.join(", "),
+                scenario.description,
+            );
         }
         return Ok(());
     }
@@ -276,41 +264,40 @@ async fn run(cli: Cli) -> Result<()> {
     let endpoint = resolve_endpoint(&cli, &dotenv_values).map_err(|msg| anyhow::anyhow!(msg))?;
     let dial = resolve_dial(&cli, &dotenv_values);
     let x_token = resolve_x_token(&cli, &dotenv_values);
+    let config = match &cli.config_file {
+        Some(path) => Config::from_file(path).context("failed to load config file")?,
+        None => Config::default(),
+    };
     let run_config = RunConfig {
         endpoint,
         dial,
         x_token,
+        config,
     };
 
     match &cli.command {
-        Commands::List => Ok(()),
-        Commands::All => {
-            let scenarios = [
-                Scenario::SysvarAccount,
-                Scenario::SysvarBlock,
-                Scenario::FullBlocks,
-                Scenario::Replay,
-                Scenario::Deshred,
-                Scenario::AnyCommitment,
-                Scenario::TokenOwnerBalanceChanged,
-                Scenario::Ordering,
-                Scenario::FilterAccounts,
-                Scenario::SlotDuplicate,
-                Scenario::ReplayOrdering,
-                Scenario::SlotStatusParentPresent,
-            ];
-
-            for scenario in scenarios {
-                log::info!("running scenario: {}", scenario.name());
-                run_scenario(&scenario, &run_config)
+        Commands::List { .. } => Ok(()),
+        Commands::All {
+            ref tags,
+            ref module,
+        } => {
+            for scenario in inventory::iter::<Scenario> {
+                if !matches_tags(scenario, tags) || !matches_module(scenario, module.as_str()) {
+                    continue;
+                }
+                log::info!("running scenario: {}", scenario.name);
+                run_scenario(scenario, &run_config)
                     .await
-                    .with_context(|| format!("scenario '{}' failed", scenario.name()))?;
+                    .with_context(|| format!("scenario '{}' failed", scenario.name))?;
             }
             Ok(())
         }
-        Commands::Run { scenario } => run_scenario(scenario, &run_config)
-            .await
-            .with_context(|| format!("scenario '{}' failed", scenario.name())),
+        Commands::Run { scenario } => {
+            let entry = find_scenario(scenario)?;
+            run_scenario(entry, &run_config)
+                .await
+                .with_context(|| format!("scenario '{}' failed", scenario))
+        }
     }
 }
